Revert "COMMERCE-4881 move"
This reverts commit dc16e9320f86211a70d05ff700268eee00fb9942.
